requirements.txt
appnope==0.1.0
backcall==0.1.0
certifi==2019.6.16
chardet==3.0.4
decorator==4.4.0
django-extensions==2.1.9
djangorestframework==3.9.4
gunicorn==19.9.0
idna==2.8
ipdb==0.12
ipython==7.6.1
ipython-genutils==0.2.0
jedi==0.14.0
parso==0.5.0
pexpect==4.7.0
phonenumbers==8.10.14
pickleshare==0.7.5
prompt-toolkit==2.0.9
psycopg2==2.8.3
ptyprocess==0.6.0
Pygments==2.4.2
PyJWT==1.7.1
PySocks==1.7.0
pytz==2019.1
requests==2.22.0
six==1.12.0
sqlparse==0.3.0
traitlets==4.3.2
twilio==6.29.0
urllib3==1.25.3
wcwidth==0.1.7
whitenoise==4.1.2

Django==2.2.3
django-heroku==0.3.1
django-redis-cache
celery==4.3.0
dj-database-url==0.5.0
django-dirtyfields==1.3.1

bcrypt==3.1.7